+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+ - Added `per_instance` for `VertexBufferSlice`.
+
 ## Version 0.9.0 (2015-09-15)
 
  - The blending, depth and stencil functions now use dedicated structs named `Blend`, `Depth` and `Stencil`.